refactor(meter-values): rename Prng.ts to PRNG.ts (issue #1936 c)
PRNG is the canonical uppercase abbreviation (Pseudo-Random Number
Generator). Uppercase acronyms in file names align with the codebase's
TypeScript PascalCase-for-modules convention (e.g. `OCPPServiceUtils.ts`).
`Prng.ts` was the outlier.

Rename performed via two-step `git mv` to survive case-insensitive
filesystems (macOS APFS default). Updates:
- Path imports (4 sites): `CoherentSampleComputer.ts`, `CoherentSession.ts`,
  `CoherentMeterValues.test.ts`, `PRNG.test.ts`.
- JSDoc `{@link}` module references (3 sites): `CoherentSession.ts` (2),
  `index.ts` (1).
- JSDoc backtick module references (2 sites): `CoherentSession.ts` (1),
  `CoherentMeterValues.test.ts` (1).
- Test suite description string: `describe('Prng', ...)` to
  `describe('PRNG', ...)` in `PRNG.test.ts`.

camelCase identifiers containing "Prng" (`voltagePrng`, `profilePickPrng`,
`socPrng`, `createStreamPrng`) preserved per TypeScript camelCase-for-
variables/functions convention.

Cosmetic; zero runtime change. Test invariant `fail: 0, skipped: 6`
preserved (2895 pass / 2901 total). Build passes.

Closes issue #1936 item (c).
