+++ src/charging-station/meter-values/index.ts
@@ -20,7 +20,7 @@
  *     module-scope runtime WeakMap (`disposeCoherentSessionRuntime`).
  *   - {@link ./CoherentMeterValueBuilder} — emit order, phase families,
  *     unit conversion, OCPP MeterValue assembly.
- *   - {@link ./Prng} — PRNG primitives (`mulberry32`, `hashLabel`,
+ *   - {@link ./PRNG} — PRNG primitives (`mulberry32`, `hashLabel`,
  *     `deriveSeed`, `createStreamPrng`).
  */
